Validation method for python benchmarks (#1152)
