+++ csrc/python_frontend/python_bindings.cpp
@@ -799,6 +799,17 @@ void initNvFuserPythonBindings(PyObject* module) {
       py::arg("dtype") = DataType::Int,
       py::return_value_policy::reference);
 
+  fusion_def.def(
+      "getValTolerances",
+      [](FusionDefinition& self, const py::iterable& input_iter) {
+        std::vector<c10::IValue> inputs;
+        for (py::handle obj : input_iter) {
+          inputs.push_back(torch::jit::toIValue(obj, c10::AnyType::get()));
+        }
+        return self.getValTolerances(inputs);
+      },
+      py::return_value_policy::reference);
+
   //! The Operators class is a nested class of FusionDefinition to allow the
   //! user to query the class for the list of operators.
   //!
